backend/requirements.txt
fastapi
uvicorn
qdrant-client
python-dotenv
openai
openai-agents
psycopg2-binary
